wip: restructure feature into clean components
