+++ apps/frontend-manage/src/components/elements/manipulation/types.ts
@@ -132,6 +132,22 @@ export interface ElementFormTypesContent extends SharedQuestionFormProps {
   type: ElementType.Content
 }
 
+export type BatchOperationActions = {
+  archive: boolean
+  unarchive: boolean
+  status?: ElementStatus
+  multiplier?: string
+  basePoints?: boolean
+}
+
+export const INITIAL_ELEMENT_BATCH_OPERATIONS: BatchOperationActions = {
+  archive: false,
+  unarchive: false,
+  status: undefined,
+  multiplier: undefined,
+  basePoints: undefined,
+}
+
 export type ElementFormTypes =
   | ElementFormTypesChoices
   | ElementFormTypesNumerical
