+++ apps/frontend-manage/src/pages/index.tsx
@@ -13,13 +13,13 @@ import { useTranslations } from 'next-intl'
 import { useRouter } from 'next/router'
 import { Suspense, useEffect, useState } from 'react'
 import { isEmpty } from 'remeda'
-import ElementBatchOperationsModal from '~/components/elements/manipulation/ElementBatchOperationsModal'
 import ActivityCreation from '../components/activities/ActivityCreation'
 import SuspendedCreationButtons from '../components/activities/creation/SuspendedCreationButtons'
 import Pagination from '../components/common/Pagination'
 import ElementList from '../components/elements/ElementList'
 import ElementListSearch from '../components/elements/ElementListSearch'
 import ElementListSorting from '../components/elements/ElementListSorting'
+import ElementBatchOperationsModal from '../components/elements/manipulation/ElementBatchOperationsModal'
 import ElementEditModal, {
   ElementEditMode,
 } from '../components/elements/manipulation/ElementEditModal'
